// src/Apis/Pokeapi.ts
import Axios from 'axios';
Axios.defaults.baseURL = 'https://pokeapi.co/api/v2/';

export const pokeApi = {
    getPokemon(pokemonName:string){
       return Axios.get<Pokemon>(`pokemon/${pokemonName}`);
    },
    getPokemonSpecies(pokemon: Pokemon){
        return Axios.get<PokemonSpecies>(pokemon.species.url);
    }
};

// Generated by https://quicktype.io

export interface Pokemon {
    abilities:                Ability[];
    base_experience:          number;
    forms:                    Species[];
    game_indices:             GameIndex[];
    height:                   number;
    held_items:               HeldItem[];
    id:                       number;
    is_default:               boolean;
    location_area_encounters: string;
    moves:                    Move[];
    name:                     string;
    order:                    number;
    species:                  Species;
    sprites:                  Sprites;
    stats:                    Stat[];
    types:                    Type[];
    weight:                   number;
}

export interface Ability {
    ability:   Species;
    is_hidden: boolean;
    slot:      number;
}

export interface Species {
    name: string;
    url:  string;
}

export interface GameIndex {
    game_index: number;
    version:    Species;
}

export interface HeldItem {
    item:            Species;
    version_details: VersionDetail[];
}

export interface VersionDetail {
    rarity:  number;
    version: Species;
}

export interface Move {
    move:                  Species;
    version_group_details: VersionGroupDetail[];
}

export interface VersionGroupDetail {
    level_learned_at:  number;
    move_learn_method: Species;
    version_group:     Species;
}

export interface Sprites {
    back_default:       string;
    back_female:        null;
    back_shiny:         string;
    back_shiny_female:  null;
    front_default:      string;
    front_female:       null;
    front_shiny:        string;
    front_shiny_female: null;
}

export interface Stat {
    base_stat: number;
    effort:    number;
    stat:      Species;
}

export interface Type {
    slot: number;
    type: Species;
}

// Generated by https://quicktype.io

export interface PokemonSpecies {
    id:                     number;
    name:                   string;
    order:                  number;
    gender_rate:            number;
    capture_rate:           number;
    base_happiness:         number;
    is_baby:                boolean;
    hatch_counter:          number;
    has_gender_differences: boolean;
    forms_switchable:       boolean;
    growth_rate:            Color;
    pokedex_numbers:        PokedexNumber[];
    egg_groups:             Color[];
    color:                  Color;
    shape:                  Color;
    evolves_from_species:   Color;
    evolution_chain:        EvolutionChain;
    habitat:                null;
    generation:             Color;
    names:                  Name[];
    flavor_text_entries:    FlavorTextEntry[];
    form_descriptions:      FormDescription[];
    genera:                 Genus[];
    varieties:              Variety[];
}

export interface Color {
    name: string;
    url:  string;
}

export interface EvolutionChain {
    url: string;
}

export interface FlavorTextEntry {
    flavor_text: string;
    language:    Color;
    version:     Color;
}

export interface FormDescription {
    description: string;
    language:    Color;
}

export interface Genus {
    genus:    string;
    language: Color;
}

export interface Name {
    name:     string;
    language: Color;
}

export interface PokedexNumber {
    entry_number: number;
    pokedex:      Color;
}

export interface Variety {
    is_default: boolean;
    pokemon:    Color;
}